Remove old code & dependencies
