+++ .github/dependabot.yml
@@ -1,100 +1,100 @@
 version: 2
 updates:
-- package-ecosystem: npm
-  directory: "/"
-  schedule:
-    interval: daily
-  open-pull-requests-limit: 10
-  ignore:
-  - dependency-name: framer-motion
-    versions:
-    - 4.1.5
-    - 4.1.6
-    - 4.1.8
-    - 4.1.9
-  - dependency-name: "@chakra-ui/icons"
-    versions:
-    - 1.0.11
-  - dependency-name: vite
-    versions:
-    - 2.2.1
-  - dependency-name: "@chakra-ui/react"
-    versions:
-    - 1.5.2
-  - dependency-name: snowpack
-    versions:
-    - 3.0.12
-    - 3.0.13
-    - 3.1.2
-    - 3.2.2
-    - 3.3.0
-  - dependency-name: "@web/test-runner"
-    versions:
-    - 0.12.14
-    - 0.12.15
-    - 0.12.16
-    - 0.12.17
-    - 0.12.18
-    - 0.12.19
-  - dependency-name: typescript
-    versions:
-    - 4.1.4
-    - 4.1.5
-    - 4.2.2
-    - 4.2.3
-    - 4.2.4
-  - dependency-name: "@types/chai"
-    versions:
-    - 4.2.15
-    - 4.2.16
-  - dependency-name: "@snowpack/plugin-svelte"
-    versions:
-    - 3.6.0
-    - 3.6.1
-  - dependency-name: svelte
-    versions:
-    - 3.32.2
-    - 3.32.3
-    - 3.34.0
-    - 3.35.0
-    - 3.36.0
-    - 3.37.0
-  - dependency-name: svelte-preprocess
-    versions:
-    - 4.6.6
-    - 4.6.7
-    - 4.6.8
-    - 4.6.9
-    - 4.7.0
-  - dependency-name: "@snowpack/web-test-runner-plugin"
-    versions:
-    - 0.2.2
-  - dependency-name: "@types/mocha"
-    versions:
-    - 8.2.1
-    - 8.2.2
-  - dependency-name: chai
-    versions:
-    - 4.3.0
-    - 4.3.1
-    - 4.3.3
-    - 4.3.4
-  - dependency-name: svelte-check
-    versions:
-    - 1.1.29
-    - 1.1.30
-    - 1.1.31
-    - 1.1.32
-  - dependency-name: sirv-cli
-    versions:
-    - 1.0.11
-  - dependency-name: "@rollup/plugin-node-resolve"
-    versions:
-    - 11.1.1
-  - dependency-name: "@rollup/plugin-typescript"
-    versions:
-    - 8.1.1
-  - dependency-name: rollup
-    versions:
-    - 2.38.1
-    - 2.38.3
+  - package-ecosystem: npm
+    directory: '/'
+    schedule:
+      interval: daily
+    open-pull-requests-limit: 10
+    ignore:
+      - dependency-name: framer-motion
+        versions:
+          - 4.1.5
+          - 4.1.6
+          - 4.1.8
+          - 4.1.9
+      - dependency-name: '@chakra-ui/icons'
+        versions:
+          - 1.0.11
+      - dependency-name: vite
+        versions:
+          - 2.2.1
+      - dependency-name: '@chakra-ui/react'
+        versions:
+          - 1.5.2
+      - dependency-name: snowpack
+        versions:
+          - 3.0.12
+          - 3.0.13
+          - 3.1.2
+          - 3.2.2
+          - 3.3.0
+      - dependency-name: '@web/test-runner'
+        versions:
+          - 0.12.14
+          - 0.12.15
+          - 0.12.16
+          - 0.12.17
+          - 0.12.18
+          - 0.12.19
+      - dependency-name: typescript
+        versions:
+          - 4.1.4
+          - 4.1.5
+          - 4.2.2
+          - 4.2.3
+          - 4.2.4
+      - dependency-name: '@types/chai'
+        versions:
+          - 4.2.15
+          - 4.2.16
+      - dependency-name: '@snowpack/plugin-svelte'
+        versions:
+          - 3.6.0
+          - 3.6.1
+      - dependency-name: svelte
+        versions:
+          - 3.32.2
+          - 3.32.3
+          - 3.34.0
+          - 3.35.0
+          - 3.36.0
+          - 3.37.0
+      - dependency-name: svelte-preprocess
+        versions:
+          - 4.6.6
+          - 4.6.7
+          - 4.6.8
+          - 4.6.9
+          - 4.7.0
+      - dependency-name: '@snowpack/web-test-runner-plugin'
+        versions:
+          - 0.2.2
+      - dependency-name: '@types/mocha'
+        versions:
+          - 8.2.1
+          - 8.2.2
+      - dependency-name: chai
+        versions:
+          - 4.3.0
+          - 4.3.1
+          - 4.3.3
+          - 4.3.4
+      - dependency-name: svelte-check
+        versions:
+          - 1.1.29
+          - 1.1.30
+          - 1.1.31
+          - 1.1.32
+      - dependency-name: sirv-cli
+        versions:
+          - 1.0.11
+      - dependency-name: '@rollup/plugin-node-resolve'
+        versions:
+          - 11.1.1
+      - dependency-name: '@rollup/plugin-typescript'
+        versions:
+          - 8.1.1
+      - dependency-name: rollup
+        versions:
+          - 2.38.1
+          - 2.38.3
