+++ index.html
@@ -13,7 +13,7 @@
 	</head>
 	<body>
 		<div id="root"></div>
-		<script type="module" src="/src/v2/main.tsx"></script>
+		<script type="module" src="/src/main.tsx"></script>
 		<script async defer src="https://sa.gqlshare.dev/latest.js"></script>
 		<noscript
 			><img
