+++ package.json
@@ -1,7 +1,11 @@
 {
-  "name": "gql-share-ts",
-  "version": "0.0.0",
+  "name": "gql-share",
+  "version": "1.0.0",
   "license": "UNLICENSED",
+  "author": {
+    "name": "Luke Martin",
+    "url": "https://luke.dev"
+  },
   "scripts": {
     "dev": "vite --port 4546",
     "build": "tsc && vite build",
@@ -10,10 +14,6 @@
     "cypress:run": "cypress run"
   },
   "dependencies": {
-    "@chakra-ui/icons": "^1.0.15",
-    "@chakra-ui/react": "^1.6.6",
-    "@emotion/react": "^11",
-    "@emotion/styled": "^11",
     "@modulz/radix-icons": "^4.0.0",
     "@radix-ui/colors": "^0.1.7",
     "@radix-ui/react-accessible-icon": "^0.0.16",
@@ -21,7 +21,6 @@
     "@radix-ui/react-visually-hidden": "^0.1.0",
     "@stitches/react": "^1.0.0",
     "copy-to-clipboard": "^3.3.1",
-    "framer-motion": "^4",
     "lz-string": "^1.4.4",
     "react": "^17.0.0",
     "react-dom": "^17.0.0"
