+++ src/main.tsx
@@ -1,15 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { ChakraProvider, ColorModeScript } from '@chakra-ui/react';
 import { App } from './App';
-import { theme } from './theme';
 
 ReactDOM.render(
 	<React.StrictMode>
-		<ChakraProvider resetCSS={true} theme={theme}>
-			<ColorModeScript initialColorMode={theme.config.initialColorMode} />
-			<App />
-		</ChakraProvider>
+		<App />
 	</React.StrictMode>,
 	document.getElementById('root'),
 );
